backend/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-environ==0.4.5
djangorestframework==3.11.0
idna==2.9
public==2019.4.13
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
values==2019.4.13
